refact!: Remove `FieldClass`
